requirements.txt
colorama==0.4.6
customtkinter==5.2.1
numpy==2.2.5
pandas==2.2.3
PyMeasure==0.14.0
PyQt5==5.15.11
PyQt5_sip==12.16.0
pyqtgraph==0.13.1
pyserial==3.5
PyVISA==1.13.0
PyYAML==6.0.2
qt_material==2.12
scipy==1.15.2
tkinterdnd2==0.3.0
